Time: 7 ms (1.95%), Space: 8.5 MB (1.62%) - LeetHub

+++ 0172-factorial-trailing-zeroes/0172-factorial-trailing-zeroes.cpp
@@ -5,7 +5,7 @@ public:
     {
         int sum = 0;
         int i = 1;
-        while ((int)(n/pow(5, i)) > 0)
+        while (n/pow(5, i) > 0)
         {
             sum += (int)(n/pow(5, i));
             i++;
